Removed limit on textures loaded from models - textures larger than 256 px wide and 256 x 256 total are now loadable. Updated version number in both OVR and MacOS versions to 1.0.5 .

+++ id1/model.h
@@ -115,8 +115,8 @@ typedef struct msurface_s
 // surface generation data
 	struct surfcache_s	*cachespots[MIPLEVELS];
 
-	short		texturemins[2];
-	short		extents[2];
+	int			texturemins[2];
+	int			extents[2];
 
 	mtexinfo_t	*texinfo;
